+++ src/api/v1/routes/authRouter.js
@@ -1,8 +1,8 @@
 const express = require('express');
-const { login } = require('../controllers/authControllers');
+const { login, checkAuthStatus } = require('../controllers/authControllers');
 
 const router = express.Router();
 
-router.route('/').post(login);
+router.route('/').get(checkAuthStatus).post(login);
 
 module.exports = router;