+++ src/api/v1/routes/hashRouter.js
@@ -1,9 +1,11 @@
 const express = require('express');
-const { createURLHash, getAllHashes, editURLHash } = require('../controllers/hashControllers');
+const {
+  createURLHash, getAllHashes, editURLHash, getHash, deleteURLHash,
+} = require('../controllers/hashControllers');
 
 const router = express.Router();
 
 router.route('/').get(getAllHashes).post(createURLHash);
-router.put('/:hash', editURLHash);
+router.route('/:hash').get(getHash).put(editURLHash).delete(deleteURLHash);
 
 module.exports = router;